refactor per review + script to start redis broker
