+++ scripts/brokers/redis.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+set -e
+
+docker run -p 6379:6379 --rm redis
