+++ src/prelude.rs
@@ -1,6 +1,5 @@
 //! A "prelude" for users of the `celery` crate.
 
-// TODO: Export RedisBroker once it is working.
-pub use crate::broker::AMQPBroker;
+pub use crate::broker::{AMQPBroker, RedisBroker};
 pub use crate::error::*;
 pub use crate::task::{Task, TaskResult, TaskResultExt};
